refactoring
https://online.sbis.ru/doc/e238a42e-fedf-4601-bb71-e77dd475cff6
